# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-garage Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://backend.egarage.store",
            "description": "Dev server online"
        },
        {
            "url": "http://localhost",
            "description": "Dev server offline"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Operates about user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Register"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Register"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reset-password/send-link": {
            "put": {
                "tags": [
                    "Reset Password"
                ],
                "summary": "Send link to user's mail via mail.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendResetPasswordLink"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/reset-password/": {
            "put": {
                "tags": [
                    "Reset Password"
                ],
                "summary": "Resets password.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Register": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "email|string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "password_confirmation": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "name": "JohnDoe",
                    "email": "cool@email.com",
                    "password": "12345678",
                    "password_confirmation": "12345678"
                }
            },
            "SendResetPasswordLink": {
                "properties": {
                    "email": {
                        "type": "email|string"
                    }
                },
                "type": "object",
                "example": {
                    "email": "cool@email.com"
                }
            },
            "ResetPassword": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "email": {
                        "type": "email|string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "password_confirmation": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "token": "cdacc0d252881102003fb486b419f2de52d8f7b1779585342e5b04c2003ce10d",
                    "email": "cool@email.com",
                    "password": "12345678",
                    "password_confirmation": "12345678"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}